scripts/requirements.txt
python-consul
flask